+++ tests/testthat/test-distinct.R
@@ -7,8 +7,8 @@ df <- data.frame(
 )
 tbls <- test_load(df)
 
-test_that("distinct equivalent to local unique", {
-  compare_tbls(tbls, function(x) x %>% distinct(), ref = unique(df))
+test_that("distinct equivalent to local unique when keep_all is TRUE", {
+  compare_tbls(tbls, function(x) x %>% distinct(x, y, z, .keep_all = TRUE), ref = unique(df))
 })
 
 test_that("distinct removes duplicates (sql)", {
@@ -17,6 +17,18 @@ test_that("distinct removes duplicates (sql)", {
 
 test_that("distinct works for 0-sized columns (#1437)", {
   df <- data_frame(x = 1:10) %>% select(-x)
-  ddf <- distinct(df)
+  ddf <- distinct(df, x)
   expect_equal( ncol(ddf), 0L )
+})
+
+test_that("distinct requires some variables", {
+  df <- data_frame(x = 1)
+  expect_error(distinct(df), "No variables selected")
+})
+
+test_that("by default distinct keeps only specified cols", {
+  df <- data_frame(x = c(1, 1, 1))
+
+  expect_equal(df %>% distinct(x), data_frame(x = 1))
+  expect_equal(df %>% group_by(x) %>% distinct(), data_frame(x = 1))
 })